+++ CHANGELOG.md
@@ -8,12 +8,27 @@ Please restart Homebridge after every plugin update.
 ## Current To-Do and In-Work List (For Future Releases, in rough order of priority):
 * Add ability to log and read current program name
 
+
+## 2.3.0-beta.8 (2024-01-25)
+* Added auto endpoint detection for all services, this fixes connection issues in many countries
+* Added ability to set authentication method. You must select the method in the plugin config. If none set, logon method falls back to using country code
+* Added Disable Session Watchdog to config.schema to make it easier to debug by turning off the session watchdog
+* Fixed issue connecting to mqtt broker (issue started ca. 23 Jan 2024) by adding extra subprotocol headers
+* Fixed bug in getMostWatchedChannels where the endpoint was incorrect
+* Updated Readme plugin status for various countries
+* Updated iOS version references in Readme
+* Bumped dependency "axios": "^1.6.6"
+* Bumped dependency "mqtt": "^5.3.5"
+* IN WORK: Reworking GB authentication methods. NOT YET WORKING, PLEASE BE PATIENT
+
+
 ## 2.2.16 (2024-01-16)
 * Removed AZ, CZ, DE, HU, RO from config.json and Readme. These countries no longer offer UPC TV.
 
+
 ## 2.2.15 (2024-01-14)
 * Fixed issue with MQTT connection failure in CH due to change of MQTT endpoint
-* Bumped dependency "axios-cookiejar-support": "^5.0.0",
+* Bumped dependency "axios-cookiejar-support": "^5.0.0"
 
 
 ## 2.2.14 (2024-01-06)